fixed numeric overflow (#1026)
git-svn-id: http://svn.code.sf.net/p/qmmp-dev/code/branches/qmmp-1.3@9107 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/qmmpui/playlistmodel.cpp
@@ -712,7 +712,7 @@ int PlayListModel::firstSelectedLower(int row)
     return -1;
 }
 
-int PlayListModel::totalDuration() const
+qint64 PlayListModel::totalDuration() const
 {
     return m_total_duration;
 }
